+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Breaking changes:
 
 - retry_policy() must be called inside a try block
+- variable declarations with `var` are not supported, only `let` and `const`
 
 ## Version 0.10.0 - 2025-04-28